backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-multipart==0.0.9
python-dotenv==1.0.1

langchain>=0.2.0,<0.3.0
langchain-openai
langchain-community
langchain-text-splitters
langgraph
langchain-chroma
openai

pypdf==4.2.0
chromadb==0.5.23
tiktoken

pydantic==2.7.4
httpx